+++ modules/registrar.go
@@ -24,7 +24,7 @@ import (
 	"github.com/forbole/bdjuno/v3/modules/distribution"
 	"github.com/forbole/bdjuno/v3/modules/feegrant"
 
-	"github.com/forbole/bdjuno/v3/modules/block"
+	"github.com/forbole/bdjuno/v3/modules/daily_refetch"
 	"github.com/forbole/bdjuno/v3/modules/gov"
 	"github.com/forbole/bdjuno/v3/modules/mint"
 	"github.com/forbole/bdjuno/v3/modules/modules"
@@ -75,8 +75,8 @@ func (r *Registrar) BuildModules(ctx registrar.Context) jmodules.Modules {
 	actionsModule := actions.NewModule(ctx.JunoConfig, ctx.EncodingConfig)
 	authModule := auth.NewModule(r.parser, cdc, db)
 	bankModule := bank.NewModule(r.parser, sources.BankSource, cdc, db)
-	blockModule := block.NewModule(ctx.Proxy, db)
 	consensusModule := consensus.NewModule(db)
+	dailyRefetchModule := daily_refetch.NewModule(ctx.Proxy, db)
 	distrModule := distribution.NewModule(sources.DistrSource, cdc, db)
 	feegrantModule := feegrant.NewModule(cdc, db)
 	mintModule := mint.NewModule(sources.MintSource, cdc, db)
@@ -92,8 +92,8 @@ func (r *Registrar) BuildModules(ctx registrar.Context) jmodules.Modules {
 		actionsModule,
 		authModule,
 		bankModule,
-		blockModule,
 		consensusModule,
+		dailyRefetchModule,
 		distrModule,
 		feegrantModule,
 		govModule,
